Stage-2-Spark/requirements.txt
pyspark==3.5.0
python-dotenv==1.0.0
py4j==0.10.9.7